v0.9

+++ lib/detail.dart
@@ -251,9 +251,31 @@ class _DetailPageState extends State<DetailPage> {
           backgroundColor: Colors.white,
           foregroundColor: Colors.grey,
           onTap: () {
-            Firestore.instance.collection('Post/'+groupENG+'/'+groupENG).document(post.postid)
-                .delete();
-            Navigator.pop(context);
+            showDialog(
+              context: context,
+              builder: (BuildContext context) {
+                // return object of type Dialog
+                return AlertDialog(
+                  title: new Text("삭제 하시겠습니까?", style: Theme.of(context).textTheme.body1, textAlign: TextAlign.center,),
+                  actions: <Widget>[
+                    // usually buttons at the bottom of the dialog
+                    new FlatButton(
+                      child: new Text("삭제", style: TextStyle(color: MainOrangeColor),),
+                      onPressed: () {
+                        Firestore.instance.collection('Post/'+groupENG+'/'+groupENG).document(post.postid)
+                            .delete();
+                        Navigator.pop(context);
+                      },
+                    ),
+                    new FlatButton(
+                      child: new Text("취소", style: TextStyle(color: GreyText1),),
+                      onPressed: () {
+                        Navigator.of(context).pop();
+                      },
+                    ),
+                  ],
+                );
+                });
           }
       ),
       SpeedDialChild(
